+++ dehaze-java/src/main/java/com/pei/dehaze/service/impl/SysUserRoleServiceImpl.java
@@ -76,7 +76,7 @@ public class SysUserRoleServiceImpl extends ServiceImpl<SysUserRoleMapper, SysUs
      */
     @Override
     public boolean hasAssignedUsers(Long roleId) {
-        int count = this.baseMapper.countUsersForRole(roleId);
+        long count = this.baseMapper.countUsersForRole(roleId);
         return count > 0;
     }
 }
